// web/src/generated/models/v1WorldEvent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type v1WorldEvent = {
    id?: string;
    type?: string;
    message?: string;
    createdAtUnixMillis?: string;
    lifeNumber?: string;
    dataJson?: string;
};